+++ lib/fuji/fujiHost.h
@@ -43,7 +43,7 @@ public:
     void set_type(fujiHostType type);
     fujiHostType get_type() { return _type; };
     bool get_menu_initialized();
-    uint8_t get_menu_entry_type(); 
+    uint16_t get_menu_entry_type(); 
 
     void set_hostname(const char *hostname);
     const char* get_hostname(char *buffer, size_t buffersize);
